Removing text based constructor

+++ ATC2027/ATC2027/ATC_Library/Altitude/Altitude.cs
@@ -27,7 +27,7 @@ namespace ATC2027.Library.Altitude
             if (cmbBoxAltitudeType.Text.ToLower() == "feet")
                 this.feet = int.Parse(text);
             else
-                this.feet = int.Parse(text) * 1000;
+                this.feet = float.Parse(text) * 1000;
         }
 
         public int GetAltitudeInFeet()
